begining of work for ble connecting the bouth halves of the keyboard
